add types directory
