+++ lib/index.ts
@@ -1,1 +1,2 @@
 export { device } from './device'
+export { theme } from './theme'
