+++ pages/_app.tsx
@@ -1,4 +1,5 @@
 import { createGlobalStyle, ThemeProvider } from 'styled-components'
+import { theme } from '../lib'
 
 // https://github.com/hankchizljaw/modern-css-reset/blob/master/dist/reset.min.css
 const GlobalStyle = createGlobalStyle`
@@ -27,15 +28,6 @@ const GlobalStyle = createGlobalStyle`
   }
 `
 
-const theme = {
-  colors: {
-    primary: '#57cc8a',
-    gray: '#757575',
-    bgDark: '#242930',
-    bgDarkGray: 'rgba(175, 186, 196, .25)'
-  },
-} as const
-
 // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
 export default function App({ Component, pageProps }) {
   return (
@@ -47,11 +39,3 @@ export default function App({ Component, pageProps }) {
     </>
   )
 }
-
-type AppTheme = typeof theme
-// NOTE: https://github.com/styled-components/styled-components/issues/1589#issuecomment-456641381
-declare module 'styled-components' {
-  // eslint-disable-next-line @typescript-eslint/no-empty-interface
-  interface DefaultTheme extends AppTheme { }
-}
-
