+++ tsconfig.json
@@ -12,7 +12,8 @@
     "moduleResolution": "node",
     "resolveJsonModule": true,
     "isolatedModules": true,
-    "jsx": "preserve"
+    "jsx": "preserve",
+    "typeRoots": ["types", "node_modules/@types"]
   },
   "exclude": [ "node_modules" ],
   "include": [ "next-env.d.ts", "**/*.ts", "**/*.tsx" ]
